Separate the point.
