+++ src/companies/factory/__init__.py
@@ -1,6 +1,7 @@
-from companies.factory.company import company, company_data, company_point, company_point_data
+from companies.factory.company import company, company_data
 from companies.factory.department import department, department_data
 from companies.factory.employee import employee_data
+from companies.factory.point import company_point, company_point_data
 from companies.factory.stock import stock, stock_data, stock_material, stock_material_data
 
 __all__ = [
