+++ src/companies/factory/company.py
@@ -1,6 +1,6 @@
 from app.testing import register
 from app.testing.types import FactoryProtocol
-from companies.models import Company, Point
+from companies.models import Company
 
 
 @register
@@ -15,19 +15,3 @@ def company_data(self: FactoryProtocol, **kwargs: dict) -> dict:
         iterations=1,
     )
     return {**schema.create()[0], **kwargs}
-
-
-@register
-def company_point_data(self: FactoryProtocol, **kwargs: dict) -> dict:
-    schema = self.schema(
-        schema=lambda: {
-            "address": self.field("address"),
-        },
-        iterations=1,
-    )
-    return {**schema.create()[0], **kwargs}
-
-
-@register
-def company_point(self: FactoryProtocol, **kwargs: dict) -> Point:
-    return self.mixer.blend(Point, **kwargs)
